Test FM Dispersion from within python.
