+++ pythonapi/test/CMakeLists.txt
@@ -2,6 +2,6 @@ add_test(NAME python_SublatticeTest COMMAND ${PYTHON_EXECUTABLE} -B ${CMAKE_CURR
 add_test(NAME python_CellTest COMMAND ${PYTHON_EXECUTABLE} -B ${CMAKE_CURRENT_SOURCE_DIR}/CellTest.py )
 add_test(NAME python_NeighborsTest COMMAND ${PYTHON_EXECUTABLE} -B ${CMAKE_CURRENT_SOURCE_DIR}/NeighborsTest.py )
 add_test(NAME python_InteractionFactoryTest COMMAND ${PYTHON_EXECUTABLE} -B ${CMAKE_CURRENT_SOURCE_DIR}/InteractionFactoryTest.py )
-add_test(NAME python_SpinWaveBuilderTest COMMAND ${PYTHON_EXECUTABLE} -B ${CMAKE_CURRENT_SOURCE_DIR}/SpinWaveBuilderTest.py )
+add_test(NAME python_FMDispersionTest COMMAND ${PYTHON_EXECUTABLE} -B ${CMAKE_CURRENT_SOURCE_DIR}/FMDispersionTest.py )
 
-set_tests_properties(python_SublatticeTest python_CellTest python_NeighborsTest python_InteractionFactoryTest python_SpinWaveBuilderTest PROPERTIES ENVIRONMENT "PYTHONPATH=${CMAKE_CURRENT_BINARY_DIR}/../")
+set_tests_properties(python_SublatticeTest python_CellTest python_NeighborsTest python_InteractionFactoryTest python_FMDispersionTest PROPERTIES ENVIRONMENT "PYTHONPATH=${CMAKE_CURRENT_BINARY_DIR}/../")
